Prohibits users that aren't logged in from run their solutions
